test: remove pre-condition that snapshot do not exists
Due to migration that runs on cluster bootstrap, sometimes we may
have snapshot after startup. So the pre-condition for this test that a
snapshot does not exist might not always satisfy.

Since the snapshot at position 0 do not actually have the state changes from
StreamProcessor, this test still has value. So we keep the test to cover such cases,
but only remove the pre-condition.