test: cover retrieveAuthorizedAuthorizationScopes
Adds a similar test as was provided for the other methods focussing on
the retrieveAuthorizedAuthorizationScopes. As this uses the same methods
under the hood, we do not expect anything new here. We're just adding
some coverage of different access patterns, given some different
authorization scenarios.

Again, some of the test scenarios overlap, but that's fine as these
tests are cheap, running fast and not depending on the implementation.

Like the other tests we again use a dedicated Scenario with builder
pattern to clarify the scenarios in detail.

(cherry picked from commit 9650a9c75a1e9236c440cd4496e87d7cbc871c7f)