feat: client-side backup list filter by prefix
This adds initial support for listing all backups that match a prefix
such as `123*` on the "client" side. The client here is the broker that
received a list of *all* backups from the backup store.

This implements the full feature but inefficiently. As a next step, we
should forward the prefix to the backup store.