feat: backup store filters backups by prefix
This optimizes listing of backups matching a checkpoint id prefix by
forwarding the prefix to the backup store. There, we can use it to
construct matching queries, thus reducing the runtime and return size.

To reduce code duplication, the logic of constructing prefixes and
regexes is moved to the API layer. Implementations may use it, as long
as they follow the common layout of storing backups under
`partitionId/checkpointId/nodeId`.