fix: chain multiple steps correctly to propagate errors
NPE was because the error was ignored in the previous
deleteBootstrapSnapshot. When the error is not null, the context
is null. This caused the NPE and hide the actual error.

To make this more testable, the logic of initializing from a
bootstrap snapshot is extracted into a utility class and added
tests for it. The test uses real implementation of snapshot
store to catch any errors due to the integration.