refactor: create and own executor service as part of store
Instead of always passing the virtual executor, just create it but leave
the possibility of injecting one for testing. This helps emphasize that
the executor is closed by the store, which could be misleading
otherwise, and is not consistent with other stores.

This also fixes some minor logging issues, like logging at info level
when we save a backup.

(cherry picked from commit 7d133948424599b4e9b3124154091dd81ffdf028)