test: add mapping rule matcher tests
This adds some basic unit tests to cover different edge cases of the
MappingRuleMatcher. It's set up as a nested test as we only test the
matchingRules method for now. If the class is expanded in the future, we
can easily support the testing of them as well.

The tests are quite straightforward otherwise, covering some basic
JSONPath concepts like nested paths, as well as error cases like invalid
expressions, and non-matching types.

Lastly, I've added a test case to cover the internal caching. It does
not actually test that the caching works as expected, but just to ensure
that some of its implementation is covered specifically in these tests.

(cherry picked from commit 965e933f79d8f71062e45c10185c2df1ec55cd81)