fix: don't validate actions on zeebe user tasks
Zeebe user tasks (or Camunda user tasks) are managed by the workflow
engine. So, the tasklist V1 API backend should not validate these
actions.

Note that the tasklist V1 API backend uses the secondary storage, which
is eventually consistent to the runtime state. Therefore, the tasklist
V1 API backend is unable to correctly validate the task. It is not
always aware of the current state of the task. It's up to the workflow
engine to do validations.