test: refactor network disconnect test
Using segment count to verify whether broker has received new
entries is fragile, because a snapshot in the meantime can compact
the logs resulting in flaky tests. Besides that this is too dependent
on the internal implementation which is not suitable for an integration
test.

To improve the test, we now verify that the reconnected node can become
leader. It can become leader again only if it has received new entries.

Besides this a unit test in raft already verifies that a reconnected node
receives new entries.

(cherry picked from commit 64ad255cb68488ef6b6475ae1436312e438c736c)