refactor: dont depend on persisted task
The assign, unassign, and complete methods all depend on the modified
TaskEntity returned from the respective taskStore methods. We want to
remove the usage of these methods for the zeebe user tasks as it's not
tasklist's responsibility to manage their state. As a first step, we'll
no longer depend on the returned object.

We can do this by copying the TaskEntity before the taskStore call, and
modifying it. The adjustments in this commit do this in the exact same
way as the returned entity that they replace. For example, the
taskStore.persistTaskClaim method makes a copy of the provided
TaskEntity, and then sets the assignee. We now do the same.