fix: no longer persist zeebe user task changes
Tasklist V1 API should not persist user task changes in the secondary
storage for Zeebe User Tasks. These are managed by the Zeebe workflow
engine, such that the secondary storage is updated by the Camunda
Exporter. This commit does this for the assign, unassign, and complete
methods.

For unassign specifically, there's also no need anymore to rollback the
change if an exception is caught.