refactor: remove duplication
After the changes there was some duplication between the logic of
assigning/completing the Zeebe and the Job-based user tasks.

We can easily extract this by producing a taskAfter TaskEntity in each
block. The duplicated logic can then done regardless of which user task
implementation this concerns.

We can apply this to the assign method and the completion method:
- after assigning, the metrics need to be tracked regardless of
  implementation
- after completion, the variables must be persisted and metrics tracked
  regardless of implementation

(cherry picked from commit 02ce9c089598088c807fe7421dd48ffc2588904e)