fix: ensure completing retry of join request to a single member cluster
This commit consists of fixes for the error scenarios identified by the
randomized test.

1. The joining member can start in a PASSIVE role to ensure it can
respond to requests from the other member. This is required when the
other member has started the re-configure process, but this member did
not receive the entries to commit before the restart.
2. If the joining member has already partially or fully committed the
configuration changes, then on restart it should start with that
configuration to ensure that it is an active member of the cluster so
that the other member can become the leader and complete the
re-configuration process.
3. The leader should identify duplicate join request instead of failing
the request. This is required so that the joining member stays active to
 complete the re-configuration steps.

(cherry picked from commit 5684ce8a3c7aef1abed4fe7893a96c6fa2943474)