ci: add running of archunit to the ci.yml (WIP)
We have added a new job to the ci.yml. The job is supposed to run the tests in the archunit package of the qa module.

A challenge here is that these archunit tests should run against the whole codebase. As a result we must build the
entire codebase (-am flag), yet we only want to run the archunit tests (-Dtest='*ArchTest'). The downside of this -Dtest
 flag is that all archunit tests must follow this naming pattern. On a positive note, this will ensure that archunit
 tests in a different module that follow this naming pattern still get run.

(cherry picked from commit 3317e9fc426aa76ba0a06357bcc1f12adbf5934f)