fix: configure search clients beans regardless of having the gateway embedded
In Production for "first tier" customers,
the gateway runs standalone in a separate deployment that can scale separately,
and the broker runs without exposing a gateway.

But we will still need the search-clients to be injected to the broker,
to be able to run batch operation items queries.

(cherry picked from commit dd154de29783ec1578db9b877984060d63a95d16)