fix: adding `~` char to ID regex
Motivation:
It is not possible to set the default ID pattern so that it fits every possible IdP,
but we can include already known chars into it.
Also, please note that users can override the default pattern if they need to.

If we can determine the format and chars used for client IDs,
then we can make that part of the default pattern (as it is done it this PR)

I am not adding extra chars in order to make the change minimal.