fix: make RoutingState backward compatible between 8.8 and 8.7
In 8.8 an incompatibility between RoutingState has been introduced by
reusing the field #2 for a new field, therefore introduce an
incompatibility between the two formats.

For compatibility, if `request_handling` is not set, but the deprecated
`activePartitions`, then we fallback to RequestHandling.AllPartitions as
dynamic scaling is not supported for 8.7.

The fallback should be removed in 8.9 when 8.8 is released.

closes #38140

(cherry picked from commit 784803ebfa25a71de815e9b391063a7edb8c37ff)