ci: run daily qa on main and all stable branches
Refactor daily QA workflow to compute stable branches rather
than relying on a specific number of officially supported versions.

This workflow could only run for main and three stable branches.
This let to a problem where not all officially supported versions
were being tested.
It achieved this by fetching the tags to find the versions and then
computing the matrix for those versions.
As 8.8 has not been released yet, it was not running QA for it.
This, while there is a stable/8.8 branch already.

Instead, we can fetch the stable branches and dynamically
compute the matrix. To achieve this, we need to rely on some
jq magic to form a json array for the cases where we know the
version explicitly (which we need for the generation template),
and add the main branch separately.

By computing the matrix in one go, we can greatly simplify
usage of the matrix as a strategy.

Lastly, this removes the reliance on specific env vars. As
mentioned already in the old code, the matrix cannot access
env vars. So there's a need to use job outputs. We can thus
completely remove the usage of env vars to simplify the
workflow.