ci: handle version edge cases
For 8.5, there is no Camunda 8.5+gen1 generation available
on INT anymore. We can pick the latest instead.

For 8.4, it is in extended support. So we no longer have to run the
QA for it on a daily basis anymore. If needed, it can be tested
using the zeebe-qa-testbench.yml workflow instead.