fix: configure ResourceAccessController beans regardless of having the gateway embedded
Similar to #38211,
In production for some customers,
the gateway runs standalone in a separate deployment that can scale separately,
and the broker runs without exposing a gateway.

But we will still need the `ResourceAccessControllers` to be injected to the broker,
to be able to authorize batch operation items queries in the client providers.

(cherry picked from commit 159b7eebbc0abed9b4e09788ee9361f1cd3d755e)