fix: Elasticsearch Memory Leak
When the elasticsearch exporter is used and if that fail to open
successfully, the threadpools elastic client created is left
dangling. Whilst the client object should get GC'ed, the underneath
http client thread is a reactor event dispatch thread which seems to
have is own stack and does not get properly cleaned unless we call
the close method. Lack of this resulted in trying to create
camunda exporters in a loop, and eventually no of threads for
elasticsearch-rest-client exceeded over 20k threads and resulted
in the app OOM'ing.

The bug that caused this is fixed in #37012 and it should not happen,
but this is adding a guard for any possible future failures.

(cherry picked from commit 42650d200f1e47da3a5dc6c8df2c9fc513cb78a4)