fix: Handle multi-tenancy unavailability in Connectors integration (#38273)
## Description

When multitenancy is enabled, the runtime may be unhealthy for a short
period of time which is not checked during the purge operation.
Additionally, requests from the Connectors container will gradually time
out more often. This is a stop-gap measure to counteract that behavior
by increasing the polling timeout to 1s.

## Related issues

fixes #37623