feat: use SseEmitter to detect client close
SseEmitter detects client close when it tries to write the response. If
it fails writing the response, IOException will be thrown. Later, we
catch the exception and cancel the handler for this request.

Please  note that jobs can still be returned to the closed connection.
That  means if the job is activated without a timeout, it might never
be activated. We might need to think of solutions where we should make
returned jobs activatable immediately without considering the job
timeout.