test: update test case ro tun for rest
IMPORTANT: job activation timeout is added to make the job activatable
again by worker "open". As mentioned in the fist commit, this is
required. Because jobs are returned to the controller, but they are not
written to the response as client is closed. For returned job to become
available again, timeout needs to expire, so worker "open" can pick it
and write to the response.