feat: introduce consistency policy API
Introduces a ConsistencyPolicy interface in the Java SDK. This allows defining how to handle eventual consistency of the API.
The policy defines what condition to what for, how long to wait at max, and how to backoff between retries.

All search, fetch, and statistics requests require defining a consistency policy before they can be executed.

The policy is handed to the CamundaFuture created by the requests. So far, it is not acted upon.

There are convenience methods to define consistency policies. The interface offers a "noWait" policy where the request won't retry.
Requests can also be executed with a default policy via convenience method.

All monorepo code using the client methods is refactored to define a policy.