fix: Creating Duplicate Beans for Tasklist & Operate
Previously these apps had their own indexes and we had to wire different
beans of same type, but writing to different indexes. Since v8.8 onwards,
both apps now read/write from same indexes. However, we are still creating
two beans of the same type, but naming them according to their usage.

This remove that distinction of the wired bean by name, instead create one
bean of the same type. The caveat is when running the app in standalone
mode where tasklist & operate would create one of the same beans without
being able to distinguish by name. In order to counter that, the
tasklist beans are annotated with an exclusion so they wouled only be
created when the "operate" profile is not running.

Another advantage of this is when running in standalone mode and deleting
a process, we were sending delete request to all the wired beans,
including duplicates. This will save additional "wasted" calls that get
made to delete the same object from the same duplicated indices.