feature: Archive Standalone Decision Instances
Whilst we are archiving decision instances attached to a process via
the linked `processInstanceKey` of the decision instances, the standalone
evaluations are never moved which can result in building completed
data in the hot indexes. This add a specialised archiver job that
would move any standalone decision instances evaluated prior to the
archive window, and move them to their dated indexes.