feat: Archive Standalone Decision Instances Repository Changes
Whilst we are archiving decision instances attached to a process via
the linked `processInstanceKey` of the decision instances, the standalone
evaluations are never moved which can result in building completed
data in the hot indexes.

Aim is to add a specialised archiver job that would move these standalone
instances to dated indexes. This add the repository code changes required
for such archiver job. The archiver will be added separately as that
cannot be backported due to the archiver refactor.