fix: resolve correct setters when deserializing records (or generally classes annotated with @ImmutableProtocol)
Since v2.11.0 of the immutables library, builders have two
initialization methods for fields of an immutable type:
* one accepting the field value directly
* one accepting a lambda builder

Since the immutable builders are used for the deserialization of the
immutable records, this causes a conflict in the setters definition
(note: the builder initialization methods are prefixed by "with*" in the
definition of ImmutableProtocol). This is solved by only considering the
non-lambda method in AnnotationIntrospector#resolveSetterConflict.