fix: check that all partitions have an up to date routing state to complete scaling
To avoid having a partition not being aware of the new partitions, check
that the routing state is up to date in all partitions.

Start from Protocol.DEPLOYMENT_PARTITION and stop with the newly started
partitions.
The requests are sent one at a time, following the partitionId order, so
that the first partition is invoked first.
If the first partition is not ready yet, requests to other partitions
are not sent.

(cherry picked from commit 9f6b517b26be74cdbe4ef1e2085eaee9196a1ee3)