ci: only add-to-project when issue labeled explicitly
Previously this workflow would add the issue to a particular
project on the labeled event if the issue already had the needed
label but the newly added label was actually a different one.

This led to annoying interactions if the issue should not be
part of a particular project, where the workflow would re-add the
issue to the project and the user then has to remove it again.

We can restrict these runs to the following:
- if event is not labeled (then it was just reopened or transferred)
- otherwise if it was explicitly labeled