feat: include ScaleIntent.STATUS to the set of whitelisted intents
Because we now send a request for each partition, we should avoiding
failing the request under high load.
If the request fails for one partition, the entire operation will be
retried, sending another N requests to the different partition,
increasing load.
Adding it to the white-listed intents avoids this situation.