feat: Archive Standalone Decision Instances (backport stable/8.7)
Whilst we are archiving decision instances attached to a process via
the linked processInstanceKey of the decision instances, the standalone
evaluations are never moved, which can result in building completed
data in the hot indexes. This adds a specialised archiver job that
would move any standalone decision instances evaluated prior to the
archive window, and move them to their dated indexes.

Also adding tests for this and a few others that were missing.

(cherry picked from commit 7d01f4173dd206232142999183db807356f3558d)