fix: opt-out of distributed auth info for migrating message subs
We are in a deeply-nested place within process instance migration and
don't have access to the original command's auth info. Instead of trying
to pass that in, opt-out of sharing auth info with the target partition.
This assumes that we don't need auth-info on the target partition and
that auditability is optional.