ci: tests need to be built instead of pulling snapshot artifacts
With maven.test.skip set, the snapshot test jar wasn't built, requiring
maven to pull a SNAPSHOT artifact from somewhere. Most of the time,
this worked and didn't cause problems because we don't actually exercise
the tests. It would fail occasionally because shortly after a release,
the SNAPSHOT artifacts for the next version weren't available yet.

We can make this more robust by just accepting that we need to build
test code too.