refactor: rename exporter to exporting config
This dynamic config will not just contain the exporters and their state
but will also determine whether exporting is paused or not.

In some places, we already used the more accurate "exporting", for
example in Javadocs and as OpenAPI component name.

BREAKING CHANGE: Buf says this change violates the
`FIELD_WIRE_COMPATIBLE_TYPE` rule. This rule is too conservative here,
renaming a message is a wire-safe change.