feat: support defining auth check for batch operation
Adds support for defining authorization checks for a batch operation command externally.
The service layer defines the UPDATE_PROCESS_INSTANCE permission for single process instance incidents resolution.
This way, the permissions are properly checked before a batch operation is created. But no batch permissions are needed.

This is batch operation-specific for now with a dedicated property on the corresponding record.
A general-purpose mechanism for records could be derived from this but is out of scope here.