feat: add exporting state to dynamic partition config
For now, this has no function, but it's at least (de-)serialized to
the protobuf representation. When this field is missing from the
serialized state, we will use `null` as a placeholder and re-encode
it as `EXPORTING_STATE_UNKNOWN`. We also use `null` when initializing
dynamic configuration for the first time. Eventually, this will be
initialized based on the actual state of exporting as it was set by
`PartitionAdminControl`.