refactor: defensive copies in `ClusterConfiguration` and nested types
This is generally a good practice because it removes one potential
source of mutability for records which are assumed to be immutable.

This was motivated by slow property tests. So far, we had to rely on
AssertJ's recursive comparison to compare two `ClusterConfiguration`
objects. This was caused by jqwik interfering by using sorted collection
types when creating `ClusterConfiguration` instances.
On deserialization, we would instead use unsorted collection types,
leading to a failed `equals` check. With defensive copies, we discard
jqwik's sorted collections and always use unsorted ones, letting the
`equals` check pass again. A regular `equals` is much faster than
recursive comparison, reducing the overall run time of
`ProtoBufSerializerPropertyTest` from ~20 seconds to ~6 seconds.