fix: fallback to V1 applier behavior if the incident contains the wrong element id
For very old incidents it could be that the incident record contains NO_CATCH_EVENT_FOUND as the element id. When
this happens we shouldn't just blindly add it to the job. In that case we must fallback to the V1 behavior and get
the corresponding element instance from state and use this instance's element id.