test: verify job element id is set to the right value
There is 2 cases. For older incident it could be that the incident contains NO_CATCH_EVENT_FOUND. In those cases we
must get the element instance from state and get the element id from here.
In the other case the incident record already contains the correct element id. In this case we can just set it
straight on the job.