fix: configuration change requests generate operations deterministically
The order of entries in the members and partitions maps is undefined,
leading to a non-deterministic order of operations.
This doesn't cause any big issues but is not ideal, especially because
it could differ to previous dry-run results.

Additionally, this caused tests that rely on the exact order to fail,
now that we take defensive copies of these collections.