test: disable raft flush to save IO
Flushing mostly helps with crash resiliency - something that we don't
need during tests because we won't recover from a crashed machine.
Without a crash, we can rely on the page cache of the kernel, and don't
need to care whether data was flushed or not.

The only exception to this would be micro-benchmarks but those should
customize carefully control the entire configuration and, if relevant,
enable flush anyway.

(cherry picked from commit ad57cea41b265dd2d78e78b6b351d7ce648c6766)