fix: reset data folder when setting up test application
The lifecycle between the extension and the test application are slightly different. Only one test application is
created per run and reruns. The beforeAll in the extension is triggered for each run and for each rerun. As a result
the application configuration doesn't get properly reset.
This generally isn't a problem, as the configuration is setup by the test case and will always be the same. However,
for the data folder it's different. If we reuse the same folder, it means we run into some data issues such as
entities managed by the EntityManager that already exist.

By setting the data folder to the default value we automatically create this data folder in a temporary directory.
The rerun would notice a temporary directory is already set an reuse the same one. If we reset the data folder to the
 default value here, it will create a new temporary directory to create it in.

 There is one downside to this approach. This prevents us from setting the data folder to a different folder in our
 test cases. Looking at usages we do not do this anywhere. I have accepted this limitation here, as it makes the
 implementation much much simpler.

(cherry picked from commit 177e07a59dde9a4fd06d518344e3057c3312c29d)