fix: snapshot for bootstrap is not reserved and can be deleted
When a snapshot for bootstrap is restored into the `snapshots/` folder,
it's flag `isBootstrap` in the metadata file is still set as "true",
because we just copy all files from the snapshot without modifying the
metadata.
Once it's been restored, it's included in the `availableSnapshots` field
of `FileBaseSnapshotStore` as a regular snapshot.
Method `isReserved()` is used only when deleting snapshots from the
available snapshot list, therefore we do not need to take into account
that flag.
Snapshot for bootstrap saved in the folder `bootstrap-snapshots/` are
deleted anyway within the transition steps of the partition.

(cherry picked from commit 013069957937dff9c56c81bd044b868eaee12687)