test: provide arbitrary sorted sets and maps
Now that we use those types in `ClusterConfiguration`, we need to tell
jqwik how to generate these types or instruct it to use the constructors
which take unsorted maps and sets and converts them to sorted ones
automatically.

Having a generic provider for all sorted maps and sets seemed like
the easier and more scalable approach.