feat: migrate global user task listeners configuration from legacy to Unified Configuration
In accordance with the new Unified Configuration approach, the global
user task listeners configuration has been moved from
zeebe.broker.experimental.engine.globalListeners.userTask to
camunda.cluster.global-listeners.user-task.

The implementation under the hood has remained the same, working as if
the legacy configuration was still present, but it is now exposed to the
user through the Unified Configuration which overrides the legacy one.

In the future code references to the legacy configuration could be
removed in favor of a "native" Unified Configuration approach, but the
user experience would not be changed.