feat: allow configuring a single contact point that resolves to multiple addresses
With this feature, we don't have to configure address of each broker as the initial contact
point. Instead, it can accept a single host name that can be resolved to multiple addresses.
For example, in a k8s deployment, we can configure the headless service name as the initial
contact point. All nodes behind the service name will be discovered. The addresses are
periodically refreshed so that any new nodes are also discovered.

There is no change in the user facing configuration. The existing configuration for
initial contact point can now accept both individual addresses (current behavior)
as well as the service names. Both can be configured at the same time.