fix: don't expose ElementInstanceNotFoundException
This exception is private, but was exposed because it wasn't caught in
all possible execution paths. This resulted in sometimes throwing
IllegalStateException and sometimes throwing
ElementInstanceNotFoundException.

Instead, we can simplify the code using a tryGetElementInstance method
which throws an IllegalStateException with the same metadata as before.

As we want to additionally share the element instance path in one of the
places, it made more sense to duplicate the method for it and expand it
with the additional information by overloading. The duplicated code
between both tryGetElementInstance methods is uncomplicated and is thus
acceptable. Trying to deduplicate it results in more complicated code.

These changes further simplify the code in the
buildElementTreePathProperties method which is also positive. By
naming the method try* it's clear that it may throw an exception.

(cherry picked from commit d6351499c29aa770e39e5d196ed999290746311c)