fix: adjust job update processor
The job update flow was broken due to a wrong change in the processor.
The RETRIES_UPDATED and TIMEOUT_UPDATED event weren't written anymore, for this reason the outcome in the second storage
was wrong. With this changes we reset the old behaviour, the JobUpdateProcessor writes the appropriate followup event
based on the changeset, the JobUpdatedApplier is removed because the job updates are handled by the specific appliers

(cherry picked from commit 63b7e4006959a86cadc60aa53b2c9aa593d7698a)