fix: do not use shared singleton for partition transition steps
It was an unwritten assumption that the partition transition steps are
singleton objects. This causes problems, when we accidentally add
mutable fields in the state that gets updated by multiple partitions
at the same time.

To prevent such bugs, we do not share the objects anymore. Instead
create it once for each partition. The steps can be shared between the
transitions with in a partition.

(cherry picked from commit 9ed472e0d89d5a7a17d9816478a6d5cf50fbe61a)