refactor: NodeIdRepository does not initialize in the constructor
Initializing the repository in a separate method enable us to remove the
field `clusterSize` in the class, as it's only needed when initializing.
This is safer, considering that the clusterSize in the configuration may
not reflect the actual size of the cluster.

The repository is initialized by the NodeIdProvider directly as well.