fix: Keycloak migration, refer to groups by name
Previously names were normalized to valid ids like done for the SaaS migration.
This is not the right approach for Self-Managed Keycloak though, as there
groups are not managed as entities within the OC. Instead the name managed in Keycloak
needs to be used as foreign key to grant groups authorizations or role/tenant membership.