fix: await for all authorizations
This unflakes the BroadcastSignalMultiplePartitionsTest
.shouldRejectBroadcastForOneUnauthorizedProcess test.

In the test it tries to create a process instance using the DEFAULT user
(which is supposed to have the admin role). However, this is rejected
because the user does not have the correct permissions, even though it's
supposed to be an admin.

The problem is that the identity setup may not yet have finished setting
up the admin role. This particularly concerns multi-partition setups.

The reason is that we didn't wait until all authorizations were created,
but only until a subset of them were. The mistake was that we considered
the number of resource types to be equivalent to the number of
authorizations, but that's not the reality. Instead, each of the default
roles has authorizations. The admin role already has an authorization
for each resource type, and so does the readonly admin. The RPA role has
just two authorizations, and so on.

We can simply check the INITIALIZED event for the number of
authorizations that had to be created.

Note that it would be better if the INITIALIZED event would follow the
full creation and distribution of all the identity entities (users,
roles, etc). For now, this just unflakes some tests.