fix: allow 0 as valid value for numJobWorkerExecutorThreads
This used to be a valid input in the previous implementation where
we reused the scheduled thread pool for everything.
The new approach uses a fixedThreadPool instead,
which throws an exception if maxThreads is <=0.
We now use 0 as a trigger to reuse the scheduled executor
for both scheduling and job handling.