test: fail tests if RecordingExported is not short-circuited
If, while trying to retrieve records from the RecordingExporter, the
wrapped AwaitingRecordIterator times out it means that it kept waiting
for new records which did not exist. This is commonly caused by the fact
that access to the records is done without any short-circuit logic (e.g.
using .limit()), with the consequence of wasting time waiting for
nothing.
Another possible reason for the time-out is if new records should have
been present but they were not due to a bug; in this case the test
should fail anyway.