refactor: job complete processor implements typed record processor
Remove usages of the CommandProcessor interface so that
it can be removed. The CommandProcessor interface is
is difficult to understand and is a frequent source of
processor-related bugs.

This commit also removes the DefaultJobPreconditionGuard
class and JobAccept interfaces as they aren't used anymore.

(cherry picked from commit b1aa019436d567669df907bab42df98ca725b384)