test: activate multi-instance ad-hoc sub-process
This adds a simple test to verify that we can activate the elements
inside of an instance of an ad-hoc sub-process that exists within a
multi-instance, i.e. a multi-instance ad-hoc sub-process.

We create an instance, wait for the three ad-hoc sub-processes are
activated, and then use the API to activate elements inside.