fix: resolve class cast exception
This resolves a ClassCastException that occurs when activating elements
inside an ad-hoc sub-process when that concerns a multi-instance ad-hoc
sub-process.

We can simply avoid the class cast exception, by using the dedicated
method that allows us to specify the expected type. This method
transparently checks whether or not this concerns the multi-instance
body or the inner instance (in this case an ad-hoc sub-process).

The exception was as follows:
  java.lang.ClassCastException: class ExecutableMultiInstanceBody cannot
  be cast to class ExecutableAdHocSubProcess.