feat: only reject events with exceeding attribute sizes
Only reject follow-up events, so commands and rejections can properly written.
Since only events should be exported and stored in secondary storage, this should be OK.
This could probably be expanded to commands as well, if needed.
For rejections, either size validation shouldn't be enabled, or only exceeding data must be removed from the rejection.
Otherwise, rejections cannot be written and we enter an exception loop.